# [PERF] PromQL: eliminate string-keyed maps in binary vector matching