@linasm 2025-10-30
> I have (belatedly) realised that you have ~8 PRs which all build on one another.

@bboreham actually there are only 5. In case number [3 (lazy hashing)](https://github.com/prometheus/prometheus/pull/17148) appears too complicated, can we focus on getting the first two across the line?
- https://github.com/prometheus/prometheus/pull/17131 (already received feedback and addressed it)
- https://github.com/prometheus/prometheus/pull/17142 (rebased on top of changes in the 1st)

Then we can decide how to proceed with 3rd, maybe we can even decide to skip it.